Give agents bounded workspace intelligence
