+++ diri/crates/diri-code-intelligence/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "diri-code-intelligence"
+version = "0.1.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+
+[dev-dependencies]
+tempfile.workspace = true
