+++ diri/crates/diri-engine/Cargo.toml
@@ -8,6 +8,7 @@ license.workspace = true
 [dependencies]
 alacritty_terminal.workspace = true
 base64.workspace = true
+diri-code-intelligence.workspace = true
 diri-proto.workspace = true
 diri-pty.workspace = true
 diri-terminal-state.workspace = true
