+++ diri/crates/dirijor-mcp/Cargo.toml
@@ -14,6 +14,7 @@ name = "dirijor"
 path = "src/bin/dirijor.rs"
 
 [dependencies]
+diri-code-intelligence.workspace = true
 diri-proto.workspace = true
 serde.workspace = true
 serde_json.workspace = true
